Continuation of shop system
Well I have been looking at it for about 2 hours now just to find a way to do that. I came up with the conclusion that for the click/hover/unhover etc features you need to check for which actors to do that. Right now only the actors that have the bpc_clickable can have that feature. But if i make it for child actors instead of static meshe(s) which is what we need I cant get access to the items info and in general class actually because I will need to get access to 2 things (1. its static mesh to add the highlight, 2. the "items" variable for "s items" so I can addd the item to players inventory. So the main problem which must be solved is just how to do that correctly and get the item for all that. I can call you to show you and explain what I mean if you want. So yeah its a simple thing to do but I havent done that before the same way exactly so yeah. After that small thing the logic to add the item to inventory etc will be easy its just this thing
I have made the shop like all we need its just this thing to fix somehow