Climbing System (Part 9)
- Added the feature for player to catch the ledge when on ground so for this I changed all the beginning logic for the catching the ledge logic to also do that not only when on air but also only when pressing it and finding a ledge to catch which has the surface/channel type we added before.

- I changed the action key with E but E isnt very convenient cause if you want to jump right you cant press d and e together
I would say space bar is the best for this. The easiest key for that
we can maybe keep it this way and I will research this and maybe try to make it the way you want it
you let the player know that he's waiting to get a command to hop instead of shimming so you press "space bar" and then regarding the key its hopping that direction
Maybe we can change the spacebar button to left shift.