Fixed Character Customization Level
Only camera rotation is left to be fixed! I tried refactoring the system in one other way but found out that this way we cant do something necessary so went back to this one and yeah basically I have to find a way to be able to set the rotation of the camera just as I want it.