Advanced Puzzle System
- Normal Puzzles (levers for doors)
- Music Puzzles
- Invisible Floor Puzzles

And all created with the most advanced and flexible way so you can easily add more puzzles and play with them such as if that music door will open this door or if it will make the floors even visible or if the switch will make them visible or open the door. Ofcourse I have them as we said, but Im pointing out how it can work!